sw floattable, nesting: fix PDF export
The problem was that the bugdoc rendered fine when opened interactively,
but not when converted to PDF via 'soffice --convert-to pdf'.

The direct cause was that the interactive case also enabled the
navigator that did one more calculation for each fly, and that fixes the
layout. The headless export to PDF has no navigator, so that didn't
happen.

The root of the problem seems to be that:

1) There is a bad left position for the inner follow fly (it's outside
   the page rectangle).

2) SwTabFrame::Split() for the outer tab normally calls
   lcl_RecalcSplitLine() which moves flys away + back.

3) Moving away happens in lcl_InvalidateLowerObjs() in the
   _bMoveObjsOutOfRange case, here we move away both fly 10 and 20 (both
  inner flys).

4) Moving back happens in
   SwToContentAnchoredObjectPosition::CalcPosition(), but that's not
   invoked for the inner follow fly.

Fix the problem by extending SwObjectFormatter::FormatObjsAtFrame_() to
be more consistent here and move back both flys.

This is related to tdf#55160.

(cherry picked from commit facdac2443d50339f81415d09c1869d19dded7bf)

Change-Id: Ia264a828f36cf2bc1d67986ab3a97d57e9d167a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156811
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>