Related: tdf#126449 sw floattable: fix bad position of in-table follow fly
The reduced bugdoc is of 4 pages, an inner floating table should be on
pages 2 -> 4 but was only on 2 -> 3.

The reason for this seems to be that we correctly split the inner table
twice, but the 3rd fly frame in the chain is positioned on page 3, and
once the anchor moves to page 4, we never re-position the fly frame. The
fly frame is normally only positioned once, since the position of the
anchor and the fly frame depend on each other, so the loop has to be
broken, and this is normally done by locking the position of the fly
after one positioning.

Fix the problem by extending the condition when to unlock the position
of anchored fly frames in SwTextFrame::MakePos(): similar to not yet
positioned fly frames, do one more cycle in case the anchor and the fly
frame are not on the same page.

This just fixes a simplified DOCX bugdoc, the full document still needs
more fixes.

(cherry picked from commit 83abd141bf41c1c8a1d4e5a894b235c842da2a07)

Conflicts:
	sw/source/core/text/frmform.cxx

Change-Id: I066c64cf4e12bf31f22218f84f0f2425256ba63e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157496
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>