tdf#126449 sw floattable: DOC import: handle inner floating table
One problem with the bugdoc is that the inner floating tables in the DOC
file stay in a single page.

Seems the usual to-para anchoring + allow-to-split logic is not used
here, because the toplevel table is handled at
SwWW8ImplReader::StartApo(), but the inner table is handled in
SwWW8ImplReader::StartTable(). Additionally, the toplevel table is
anchored to-para (which seems to be the closest to Word's "position this
table based on the next paragraph" concept), but the inner table was
anchored to-char, and such fly frames can't split.

Fix the problem by switching to to-para anchoring even for inner
floating tables. This improves consistency with toplevel floatint tables
from DOC and all floating tables from DOCX. It seems to the to-char
anchor type was added in commit 10f352d2faf6a4d72337b2c098a65377eee5138b
(INTEGRATION: CWS swqbugfixes18 (1.111.60); FILE MERGED, 2005-03-30),
but there the motivation was to make sure these are not inline; so that
use-case keeps working.

This does fix the overlapping text with the original bugdoc, but
otherwise the DOCX version is still slightly closer to the Word render
result.

(cherry picked from commit 89a75cd194371002247d0138e759835bc673f7b0)

Change-Id: I379a26194da4d3a06241aa3c6f5ae78606f8fc12
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157598
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>