tdf#155682 sw floattable: fix DOCX with big pictures causes endless loop
The bugdoc had a floating table in the header that overlapped with
full-page, as-char anchored images in the body text, leading to a layout
loop.

The trouble is that the body text was broken into 2 lines: the first
line just had fly portions for the area where the in-header floating
table was rendered and then the second line hosted the actual image. But
then this image didn't fit the remaining space, so it moves to the next
page. And this happened again and again.

Fix the problem by keeping in-header anchored floating tables inside the
boundary of the header, this way the body text still doesn't overlap
with the floating table, but the space available to the body text
doesn't have to contain fly portions. Which means we don't try to move
the image to a next page, so there is no loop.

Note that the IsFollowingTextFlow flag is already ignored in some cases
(e.g. sw/qa/extras/ww8export/data/tdf128700_relativeTableWidth.doc), so
this doesn't break the use-case when a footer provides an anchor
position for some text on the left/right margin. In that case the footer
height is still unchanged as it should be.

(cherry picked from commit 9704f61982360ce35983a61cca3fd00bbdf51ab6)

Change-Id: Id9e80140af3123d52b0fea2f96fc19c150c8e736
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158482
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>